Update main.js
on a scale from 1 to 5, how comfortable were you with this assignment? 4
What was a win you had with this assignment? i fully understood JQuery
What was a challenge you had with this assignment? question 10
Is there anything that you'd like some further information on? no
Other stuff on your mind? work with anyone? tell us about it!